Fix splashscreen no effect when missing to set "splashscreen".
In the CordovaActivity.java, it requires to set "splashscreen"
property for customized splashscreen when application want to
leverage cordova's splashscreen API. The "splashscreen" property
is set to 0 by default.

But from the android official document for  setBackgroundResource API
(http://developer.android.com/reference/android/view/View.html),
it shows "The resource should refer to a Drawable object or 0 to remove the background."
Which means there would be no splashscreen effect even there do
exist splashdialog, even splashscreen API has been called.

The activity of cordova-mobile-spec is auto generated from the bash, which forget
to set the "splashscreen", but it calls the cordova's splashscreen API.

This patch is to add check before setBackgroundResource API, if its argument
is euqal to 0, would skip this call this API.

BUG=https://github.com/otcshare/cordova-xwalk-android/issues/12