Reschedule the loadurl on main thread when it's run on other threads, which is not maintained by XWalk.
Because XWalk depends on content component, which owns notification service for each thread,
Each navigation need happen on the XWalk thread that's responsible for notification service
object's creation/release, otherwise, it would crash the XWalk. Such as: this function is called
by the instrumentation test, which is run in the android.test.InstrumentationTestRunner.

So when the above situation happens, cordova container needs to schedule this function to run
on the main thread, not the instrumentation thread.

BUG=https://github.com/otcshare/cordova-xwalk-android/issues/2