Align SSL error behavior with Cordova-Android.
Cordova-Android ignores all SSL errors for debug builds, and fails on
SSL errors in release builds.

Crosswalk-Cordova-Android shows an alert dialog that lets the user
ignore the SSL error or cancel the request. This behavior is surprising
to Web platform developers, as typical browser behavior matches that of
Cordova-Android for release builds.

This change implements the Cordova-Android behavior, using the recently
introduced XWalkResourceClient::onReceivedSslError() hook.