Add the option of --xwalk-shared-library to use Crosswalk shared mode
Use the option of --xwalk-shared-library to create a shared mode Crosswalk
Cordova project. It isn't used together with --shared.
Sunlin comments:
wqIn new implementation, there are 2 parallel directories named
xwalk_share_library and xwalk_core_library, we plan to release both of them in
crosswalk build. The new option is to help developers of Embedding API or Cordova
to choose which library to link when  creating their project. Regardless which library
used in built time, the final crosswalk app could check whether it is in shared mode
and check compatibility on run time.