Fix the Cordova app exit issue.
This patch is to fix Cordova app exit issue when the soft keyboard
displaying and press back key.
If the back key was not consumed by CordovaWebView, finish the current
activity, The key up in CordovaWebView was always called with soft keyboard
displaying or not, so adjust the finish() action to CordovaActivity's
onKeyDown().

BUG=XWALK-3915